requirements/prod.txt
-r base.txt
gunicorn==20.1.0
django-heroku==0.3.1
django-environ==0.4.5